requirements.txt
boto3==1.10.0
botocore==1.13.0
django-environ==0.4.5
docutils==0.15.2
isort==4.3.21
jmespath==0.9.4
python-dateutil==2.8.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.6
